[OGUI-1216] Extract zoom into its own model
Moves all state, properties and methods into own observable that bubbles up to Model.
